cambios postProducts (modularización) 06/01 11:33
