+++ server/src/routes/products.routes.js
@@ -1,9 +1,11 @@
 const { Router } = require("express");
 const router = Router();
-const { getAllProducts, postProduct } = require("./controllers/products.js");
+const { getAllProducts } = require("./controllers/getProducts.js");
+const { postProduct } = require("../routes/controllers/postProducts.js");
 
 router.get("/", getAllProducts);
 router.get("/:id", getAllProducts);
 router.get("/?name", getAllProducts);
+
 router.post('/', postProduct);
 module.exports = router;
